# Optimize batch RUT formatting and streaming paths