## Summary
- cache Rut instances inside `ResultadoLote` so batch formatting reuses validated objects instead of re-parsing strings
- streamline streaming helpers to validate with a shared `ValidadorRut` and reuse freshly built Rut instances for formatting output
- extend unit coverage to assert the cached Rut objects are exposed for downstream consumers

## Testing
- `pytest`

## Benchmarks
- `python benchmarks/formateo_batch.py` *(synthetic script used inline)*
  - validar_lista_ruts: **0.0380s → 0.0413s** (≈ +8.7% from extra bookkeeping)
  - formatear_lista_ruts: **0.0884s → 0.0517s** (≈ −41.5% due to Rut reuse)


------
https://chatgpt.com/codex/tasks/task_e_68d442cc53ac8328bc7a3a4e0d199160